pyproject.toml
[project]
name = "tissue-map-tools"
version = "0.1.0"
description = "Utilities for working with tissue maps and 3D spatial omics in Python, with a focus on mesh processing and visualization."
readme = "README.md"
authors = [
    { name = "Eric Moerth", email = "ericmoerth@g.harvard.edu" },
    { name = "Luca Marconato", email = "luca.marconato@embl.de" }
]
requires-python = ">=3.11,<3.13"
dependencies = [
    "zmesh>=1.0.0",
    "trimesh",
    "pyfqmr",
    "anywidget>=0.9.18",
    "connected-components-3d>=3.23.0",
    "fast-simplification>=0.1.9",
    "ipywidgets>=8.1.6",
    "jupyterlab>=3",
    "meshio>=5.3.5",
    "ome-types>=0.6.0",
    "ome-zarr==0.10.3",
    "open3d>=0.19.0",
    "pymeshfix>=0.17.0",
    "pyvista[all,trame]>=0.44.2",
    "scikit-learn>=1.6.1",
    "tifffile>=2025.3.30",
    "trame-jupyter-extension>=2.1.4",
    "watchfiles>=1.0.5",
    "zarr>=2.18.7",
    "cloud-volume>=12.3.1",
    "xarray>=2024.9.0",
    "vedo>=2025.5.4",
]

[dependency-groups]
extra = [
    "igneous-pipeline>=4.30.0",
    "task-queue>=2.14.2",
]
examples = [
    "dirhash>=0.5.0",
    "neuroglancer>=2.40.1",
    "napari-spatialdata>=0.5.5",
    "spatialdata>=0.4.0",
    "napari>=0.6.1",
    "devtools>=0.12.2",
    "pyqt5>=5.15.11",
    "fiona>=1.10.1",
]
dev = [
    "pre-commit>=4.2.0",
]

[tool.uv]
default-groups = ["examples", "dev"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
